+++ docs/.vitepress/components/ClassConfig.vue
@@ -8,7 +8,6 @@ import {
 import {
   computed,
   reactive,
-  ref,
 } from 'vue'
 
 interface VariantModel {
@@ -51,32 +50,11 @@ for (const variant in variants) {
     options,
   }
 }
-
-const isDark = ref<boolean>(document.documentElement.classList.contains('dark'))
-
-const observer = new MutationObserver((mutations) => {
-  for (const mutation of mutations) {
-    if (mutation.type === 'attributes' && mutation.attributeName === 'class') {
-      const darkClass = document.documentElement.classList.contains('dark')
-
-      if (darkClass !== isDark.value) {
-        isDark.value = darkClass
-      }
-    }
-  }
-})
-
-observer.observe(document.documentElement, {
-  attributes: true,
-  attributeFilter: [
-    'class',
-  ],
-})
 </script>
 
 <template>
   <VcConfigProvider locale="en-NL">
-    <VcThemeProvider :appearance="isDark ? 'dark' : 'light'">
+    <VcThemeProvider>
       <h2 id="default-styles">
         Default styles
       </h2>