+++ packages/query/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @wisemen/vue-core-query
 
+## 0.0.46
+
+### Patch Changes
+
+- Updated dependencies [79ec7a5]
+- Updated dependencies [d7e67cd]
+- Updated dependencies [566f122]
+- Updated dependencies [0726edd]
+- Updated dependencies [2cac603]
+- Updated dependencies [9e7ff47]
+- Updated dependencies [dad400c]
+  - @wisemen/vue-core-components@1.16.0
+
 ## 0.0.45
 
 ### Patch Changes